sga_backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
mysqlclient
django_extensions
django-cors-headers 